fs: aio: Optimization for SSD-only machines If the system is running on NAND-based storage, without mechanical drives, we don't need to manage the plug/unplug of the actual block because we don't have to count the time for the head to reach that disk physical location. Solid state storage have no head to move and no location to "physically reach", as the block is being retrieved in the exact moment we request it.
The do_io_submit function expects the storage device to
be ready to deliver the block (aka - next block of the
queue to process), therefore it's safe to remove useless
waits on systems using only that type of storage.

This patch adds a configuration option to enable the
SSD-only optimization for the Linux Asynchronous IO
framework.
Enabling it on systems with rotational storage (or mixed)
WILL hit storage performance A LOT.