ARM: fix backtrace Recent kernels have changed their behaviour to be more inconsistent when handling printk continuations.  With todays kernels, the output looks sane on the console, but dmesg splits individual printk()s which do not have the KERN_CONT prefix into separate lines.
Since the assembly code is not trivial to add the KERN_CONT, and we
ideally want to avoid using KERN_CONT (as multiple printk()s can race
between different threads), convert the assembly dumping the register
values to C code, and have the C code build the output a line at a
time before dumping to the console.

This avoids the KERN_CONT issue, and also avoids situations where the
output is intermixed with other console activity.

Signed-off-by: Russell King <rmk+kernel@armlinux.org.uk>
---
Noticed due to a NFS bug in 4.9-rc.

 arch/arm/kernel/traps.c  | 20 ++++++++++++++++++++
 arch/arm/lib/backtrace.S | 37 +++----------------------------------
 2 files changed, 23 insertions(+), 34 deletions(-)